gs_requirements.txt
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
constantly==23.10.4
decorator==5.1.1
docutils==0.21.2
executing==2.1.0
future==1.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==8.27.0
jedi==0.19.1
lockfile==0.12.2
lxml==5.2.1
matplotlib-inline==0.1.7
MAVProxy==1.8.70
monotonic==1.6
msgpack==1.1.0
numpy==2.1.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pymavlink==2.4.42
pyserial==3.5
python-daemon==3.0.1
PyYAML==6.0.1
setuptools==75.3.0
simplejpeg==1.7.6
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
Twisted==24.10.0
typing_extensions==4.12.2
wcwidth==0.2.13
zope.interface==7.1.1